src/main/java/ru/job4j/todo/controller/UserController.java
package ru.job4j.todo.controller;

import lombok.AllArgsConstructor;
import org.springframework.stereotype.Controller;
import org.springframework.ui.Model;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.RequestMapping;
import ru.job4j.todo.model.User;

@Controller
@RequestMapping("/users")
@AllArgsConstructor
public class UserController {
    @GetMapping("/registration")
    public String getRegistrationPage(Model model) {
        model.addAttribute("user", new User());
        return "users/registration";
    }
}
